requirements.txt
discord.py==2.3.2
Pillow==10.0.0
python-dotenv==1.0.0
aiohttp==3.8.5
openai==0.28.0
yt-dlp==2023.10.13